requirements.txt
Django>=5.0.1,<6.0
djangorestframework>=3.14.0
django-jazzmin==3.0.1
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv==1.0.1
